+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/iexec-blockchain-adapter-api/releases/tag/vNEXT) 2024
 
+### Bug Fixes
+
+- Allow null callback data when finalizing a task. (#158)
+
 ### Quality
 
 - Use poco-chain with poco v5.5.0 and voucher v1.0.0 in tests. (#152)